Change to Rt calculation to fix discretisation error
Investigation showed that the high rate of cases in the UK in January
caused a significant discretisation error between the chain binomial
model and continuous time analogues.  The previous Rt calculation
made use of a continuous time approximation, and so gave biased
estimates of Rt.

The new implementation calculates Rt using expectations of the chain
binomial model in terms of transition probabilities, not rates as
previously.